+++ src/main/scala/millfork/output/Z80Assembler.scala
@@ -73,7 +73,8 @@ class Z80Assembler(program: Program,
 
     try { instr match {
       case ZLine0(LABEL, NoRegisters, MemoryAddressConstant(Label(labelName))) =>
-        labelMap(labelName) = index
+        val bank0 = mem.banks(bank)
+        labelMap(labelName) = bank0.index -> index
         index
       case ZLine0(BYTE, NoRegisters, param) =>
         writeByte(bank, index, param)
@@ -688,7 +689,7 @@ class Z80Assembler(program: Program,
     }
   }
 
-  override def injectLabels(labelMap: Map[String, Int], code: List[ZLine]): List[ZLine] = code // TODO
+  override def injectLabels(labelMap: Map[String, (Int, Int)], code: List[ZLine]): List[ZLine] = code // TODO
 
   override def quickSimplify(code: List[ZLine]): List[ZLine] = code.map(a => a.copy(parameter = a.parameter.quickSimplify))