+++ src/main/scala/millfork/output/Z80ToX86Crossassembler.scala
@@ -22,7 +22,7 @@ class Z80ToX86Crossassembler(program: Program,
     } else code
   }
 
-  override def injectLabels(labelMap: Map[String, Int], code: List[ZLine]): List[ZLine] = code // TODO
+  override def injectLabels(labelMap: Map[String, (Int, Int)], code: List[ZLine]): List[ZLine] = code // TODO
 
   override def quickSimplify(code: List[ZLine]): List[ZLine] = code.map(a => a.copy(parameter = a.parameter.quickSimplify))
 
@@ -74,7 +74,8 @@ class Z80ToX86Crossassembler(program: Program,
     import Z80ToX86Crossassembler._
     instr match {
       case ZLine0(LABEL, NoRegisters, MemoryAddressConstant(Label(labelName))) =>
-        labelMap(labelName) = index
+        val bank0 = mem.banks(bank)
+        labelMap(labelName) = bank0.index -> index
         index
       case ZLine0(BYTE, NoRegisters, param) =>
         writeByte(bank, index, param)